More improvements to database updater
