+++ server/src/models/aircraft.js
@@ -1,7 +1,7 @@
 import cheerio from 'cheerio';
 import { model, Schema } from 'mongoose';
 
-import { getText, parseWikipediaData } from '../db/parseData';
+import { getText, parseWikipediaData } from '../db/parse';
 
 export const AircraftSchema = new Schema({
   _id: String,
@@ -10,17 +10,15 @@ export const AircraftSchema = new Schema({
   names: [String],
 });
 
-AircraftSchema.pre('save', next => {
-  this._id = `${this.iata}_${this.icao}`; // eslint-disable-line no-underscore-dangle
-  next();
-});
-
 AircraftSchema.static(
   'dataUrl',
   'https://en.wikipedia.org/wiki/List_of_aircraft_type_designators',
 );
 
-AircraftSchema.static('parseData', parseWikipediaData);
+AircraftSchema.static('parseData', data => {
+  const $ = parseWikipediaData(data);
+  return $('.wikitable tr').toArray();
+});
 
 AircraftSchema.static('getUpdate', item => {
   const $ = cheerio.load(item);
@@ -32,13 +30,14 @@ AircraftSchema.static('getUpdate', item => {
     return null;
   }
 
+  const _id = `${iata}_${icao}`; // eslint-disable-line no-underscore-dangle
   const names = tds
     .eq(2)
     .children('a')
     .map((j, a) => $(a).text())
     .get();
 
-  return { iata, icao, names };
+  return { _id, iata, icao, names };
 });
 
 export default model('Aircraft', AircraftSchema, 'aircraft');
